+++ layrzsdk/entities/__init__.py
@@ -1,24 +1,27 @@
 """ Init file """
-# Basic entities
-from .asset import Asset
-from .sensor import Sensor
-from .asset_operation_mode import AssetOperationMode
-from .device import Device
-from .custom_field import CustomField
-from .message import Message
-from .position import Position
-from .event import Event
-from .trigger import Trigger
-from .case import Case, CaseStatus
-from .user import User
-from .comment import Comment
-from .checkpoint import Checkpoint
-from .geofence import Geofence
-from .waypoint import Waypoint
-from .text_align import TextAlignment
+# Formatting entities
+from .formatting import *
 
 # Charts entities
 from .charts import *
 
+# Checkpoints entities
+from .checkpoints import *
+
 # Reports entities
 from .reports import *
+
+# Cases entitites
+from .cases import *
+
+# Events entities
+from .events import *
+
+# Telemetry entities
+from .telemetry import *
+
+# General entities
+from .general import *
+
+# Broadcast entities
+from .broadcasts import *