+++ layrzsdk/entities/checkpoints/__init__.py
@@ -1,0 +1,5 @@
+""" Init file """
+from .checkpoint import Checkpoint
+from .comment import Comment
+from .waypoint import Waypoint
+from .geofence import Geofence