+++ layrzsdk/entities/reports/col.py
@@ -1,6 +1,6 @@
 """ Report col """
 from enum import Enum
-from ..text_align import TextAlignment
+from ..formatting.text_align import TextAlignment
 
 class ReportDataType(Enum):
   """