+++ layrzsdk/entities/reports/__init__.py
@@ -3,5 +3,5 @@ from .format import ReportFormat
 from .header import ReportHeader
 from .page import ReportPage
 from .row import ReportRow
-from .col import ReportCol
+from .col import ReportCol, ReportDataType
 from .report import Report