+++ layrzsdk/entities/reports/row.py
@@ -1,5 +1,5 @@
 """ Report row """
-from ..text_align import TextAlignment
+from ..formatting.text_align import TextAlignment
 
 class ReportRow:
   """