+++ CHANGELOG.rst
@@ -4,7 +4,11 @@ Changelog
 
 v1.2.2
 ~~~~~~
-* Updated Report Col entity to set new default values
+* Updated ReportCol entity to set new default values
+* New entity ReportDataType
+* Possibility to export directly to the Report class
+* Re-organized entities/ folder
+* Added Broadcasts entities 
 
 v1.2.1
 ~~~~~~